+++ src/consts.js
@@ -9,6 +9,7 @@ const DATE_TYPE = 'date'
 const FILE_TYPE = 'file'
 const RADIO_TYPE = 'radio'
 const SELECT_TYPE = 'select'
+const CREATABLE_STRING_SELECT_TYPE = 'creatable_string_select'
 const CHECKBOX_TYPE = 'checkbox'
 const BOOLEAN_TYPE = 'boolean'
 const CURRENCY_TYPE = 'currency'
@@ -29,6 +30,7 @@ export const inputTypes = {
   SELECT_TYPE,
   RELATIONSHIP_SINGLE: 'relSingle',
   RELATIONSHIP_MULTIPLE: 'relMultiple',
+  CREATABLE_STRING_SELECT_TYPE,
   PASSWORD_TYPE,
   RADIO_TYPE,
   CHECKBOX_TYPE // not used